Replace participant actions-creators. Set default value after joining
