+++ app/game-master/action-creators/participant.js
@@ -1,5 +1,11 @@
 import { createAction } from 'redux-actions';
 
 import { PARTICIPANT_INFO } from '../actions/participant';
+import {
+    PARTICIPANT_JOINED,
+    PARTICIPANT_LEFT,
+} from '../actions/participant';
 
+export const addNewParticipant = createAction(PARTICIPANT_JOINED);
+export const removeParticipant = createAction(PARTICIPANT_LEFT);
 export const updateParticipantInfo = createAction(PARTICIPANT_INFO);
