+++ app/game-master/action-creators/round.js
@@ -11,3 +11,5 @@ export const updateRemaining = createAction(RoundActions.REMAINING);
 export const selectRound = createAction(RoundActions.SELECTED_ROUND);
 export const startRound = createAction(RoundActions.START_ROUND);
 export const stopRound = createAction(RoundActions.STOP_ROUND);
+
+export const updateParticipantSolution = createAction(RoundActions.PARTICIPANT_SOLUTION);
