+++ app/game-master/actions/participant.js
@@ -1,1 +1,3 @@
 export const PARTICIPANT_INFO = 'PARTICIPANT_INFO';
+export const PARTICIPANT_JOINED = 'PARTICIPANT_JOINED';
+export const PARTICIPANT_LEFT = 'PARTICIPANT_LEFT';
