+++ app/game-master/actions/round.js
@@ -7,3 +7,5 @@ export const REMAINING = 'REMAINING';
 export const SELECTED_ROUND = 'SELECTED_ROUND';
 export const START_ROUND = 'START_ROUND';
 export const STOP_ROUND = 'STOP_ROUND';
+
+export const PARTICIPANT_SOLUTION = 'PARTICIPANT_SOLUTION';
