--- app/game-master/actions/score.js
@@ -1,3 +1,0 @@
-export const PARTICIPANT_JOINED = 'PARTICIPANT_JOINED';
-export const PARTICIPANT_LEFT = 'PARTICIPANT_LEFT';
-export const PARTICIPANT_SOLUTION = 'PARTICIPANT_SOLUTION';
