+++ app/game-master/reducers/score.js
@@ -2,24 +2,35 @@ import MockData from '../mock-data/score';
 import {
     PARTICIPANT_JOINED,
     PARTICIPANT_LEFT,
-    PARTICIPANT_SOLUTION
-} from '../actions/score';
+} from '../actions/participant';
+import { PARTICIPANT_SOLUTION } from '../actions/round';
 
 const defaultState = MockData;
+const defaultParticipant = {
+    time: null,
+    length: 0,
+    correct: false,
+};
 
 export default function participant(state = defaultState, action) {
     switch (action.type) {
-        case PARTICIPANT_JOINED:
+        case PARTICIPANT_JOINED: {
+            const newParticipant = Object.assign(defaultParticipant, action.payload);
+
             return Object.assign({}, state, {
-                round: state.round.concat([action.payload]),
+                round: state.round.concat([newParticipant]),
+                aggregate: state.aggregate.concat([newParticipant]),
             });
-        case PARTICIPANT_LEFT:
+        }
+        case PARTICIPANT_LEFT: {
+            const filterParticipants = participant => participant.participantId !== action.payload.participantId;
+
             return Object.assign({}, state, {
-                round: state.round.filter((participant) => {
-                    return participant.participantId !== action.payload.participantId;
-                }),
+                round: state.round.filter(filterParticipants),
+                aggregate: state.aggregate.filter(filterParticipants)
             });
-        case PARTICIPANT_SOLUTION:
+        }
+        case PARTICIPANT_SOLUTION: {
             return Object.assign({}, state, {
                 round: state.round.map((participant) => {
                     if (participant.participantId === action.payload.participantId) {
@@ -29,6 +40,8 @@ export default function participant(state = defaultState, action) {
                     return participant;
                 })
             });
+        }
+
         default:
             return state;
     }
